Add files via upload
"3.19 Write a C program called time.c that determines the amount of time necessary to run a command from the command line. This program will be run as “./time ” and will report the amount of elapsed time to run the specified command. This will involve using fork() and exec() functions, as well as the gettimeofday() function to determine the elapsed time. It will also require the use of two different IPC mechanisms." (Silberschatz, Abraham, et al. Operating System Concepts. Wiley, 2018.)